requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
joblib==1.0.1
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
pydantic==1.8.1
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.22.2.post1
scipy==1.6.2
six==1.15.0
sklearn==0.0
starlette==0.13.6
threadpoolctl==2.1.0
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
